Refactoring of SPIRVReader.cpp and SPIRVWriter.cpp
Moving declarations of SPIRVWriter and LLVMToSPIRVDbgTran to SPIRVWriter.h
Moving declarations of SPIRVReader and SPIRVToLLVMDbgTran to SPIRVReader.h
Minor fixes.
