+++ lib/SPIRV/SPIRVRegularizeLLVM.cpp
@@ -125,7 +125,8 @@ bool SPIRVRegularizeLLVM::regularize() {
       for (auto II = BI->begin(), IE = BI->end(); II != IE; ++II) {
         if (auto Call = dyn_cast<CallInst>(II)) {
           Call->setTailCall(false);
-          if (Call->getCalledFunction()->isIntrinsic())
+          Function *CF = Call->getCalledFunction();
+          if (CF && CF->isIntrinsic())
             removeFnAttr(Context, Call, Attribute::NoUnwind);
         }
